Minor update to the c library (data type, README, progress percentage calculation)

Former-commit-id: e952f17395f6f279042a034b3576d209c8d16cd6 [formerly 0af7ec2d5fab913b514dcd872384c9c3dda736bd]
Former-commit-id: 9321272c96a2c58c4a304cb3f44e4fa7cad00584

+++ gravity_plot/integrator_rk_embedded.py
@@ -101,7 +101,7 @@ class RK_EMBEDDED:
 
             t = ctypes.c_double(0.0)
             dt = ctypes.c_double(dt)
-            count = ctypes.c_int(0)
+            count = ctypes.c_int64(0)
             store_count = ctypes.c_int(0)
             while t.value <= tf:
                 rk_flag = self.c_lib.rk_embedded(
